Shadow mapping
I implemented a Shadow mapping algorithm for displaying shadows in the scene. Now it is coded in the main.cpp file, but next step is to move everything where it should be. I am making this commit because it took a lot of time to make this work and don't want to lose it.