[Rendering] First part of omnidirectional lights.
Added the first steps to shadow mapping using cubemaps to render shadows out of point lights.
The shader loader code now supports geometry shaders. The meshes have the cubemap depth shaders to use depth testing with cubemaps.

Everything left is to try render the scene from a point light and generate a cubemap, and then to pass that cubemap to a light pass shader so it can adjust the light values in each mesh fragment.