[Rendering] Fourth part of omnidirectional lights
I keep trying to render the cubemap texture into its own tab in the editor. I managed to draw a red subimage in the bottom left corner of the image, which means that the display part is working fine, and it's the texture whats wrong. I'll need to check how that texture is being generated.